app/requirements-av.txt
boto3==1.21.29
botocore==1.24.30
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
easydict==1.9
fonttools==4.29.1
idna==3.3
imageio==2.16.0
jmespath==1.0.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
PyMySQL==1.0.2
pyparsing==3.0.7
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.2
PyWavelets==1.2.0
requests==2.26.0
s3transfer==0.5.2
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.2.9
torch
torchvision
tqdm==4.62.3
typing_extensions==4.1.1
urllib3==1.26.7
